+++ dataprep/clean/clean_date.py
@@ -14,537 +14,13 @@ import numpy as np
 import pandas as pd
 
 from .utils import NULL_VALUES, create_report, to_dask
-
-
-JUMP = [" ", ".", ",", ";", "-", "/", "'",
-        "st", "nd", "rd", "th",
-        "at", "on", "and", "ad", "AD", "of"]
-        #"T", "t"
-WEEKDAYS = {"Mon": 1, "Monday": 1,
-            "Tue": 2, "Tuesday": 2,
-            "Wed": 3, "Wednesday": 3,
-            "Thu" :4, "Thursday" :4,
-            "Fri" :5, "Friday" :5,
-            "Sat": 6, "Saturday" :6,
-            "Sun": 7, "Sunday":7}
-MONTHS = {"Jan": 1, "January": 1,
-          "Feb": 2, "February":2,
-          "Mar": 3, "March": 3,
-          "Apr": 4, "April": 4,
-          "May": 5,
-          "Jun": 6, "June": 6,
-          "Jul": 7, "July": 7,
-          "Aug": 8, "August": 8,
-          "Sep": 9, "Sept": 9, "September": 9,
-          "Oct": 10, "October": 10,
-          "Nov": 11, "November": 11,
-          "Dec": 12, "December" : 12}
-HMS = [("h", "hour", "hours"), ("m", "minute", "minutes"), ("s", "second", "seconds")]
-AM = ["AM", "am", "a"]
-PM = ["PM", "pm", "p"]
-ZONE = {'UTC': 0,
-        'ACT': -5,
-        'ADT': -3,
-        'AEDT': 11,
-        'AEST': 10,
-        'AKDT': -8,
-        'AKST': -9,
-        'AMST': -3,
-        'AMT': -4,
-        'ART': -3,
-        'ArabiaST': 3,
-        'AtlanticST': -4,
-        'AWST': 8,
-        'AZOST': 0,
-        'AZOT': 0,
-        'BOT': -4,
-        'BRST': -2,
-        'BRT': -3,
-        'BST': 1,
-        'BTT': 6,
-        'CAT': 2,
-        'CDT': -5,
-        'CEST': 2,
-        'CET': 1,
-        'CHOST': 9,
-        'CHOT': 8,
-        'CHUT': 10,
-        'CKT': -10,
-        'CLST': -3,
-        'CLT': -4,
-        'CentralST': -6,
-        'ChinaST': 8,
-        'CubaST': -5,
-        'ChST': 10,
-        'EASST': -5,
-        'EAST': -6,
-        'EAT':3,
-        'ECT':-5,
-        'EDT': -4,
-        'EEST': 3,
-        'EET': 2,
-        'EST': -5,
-        'FKST': -3,
-        'GFT': -3,
-        'GILT':12,
-        'GMT':0,
-        'GST':4,
-        'HKT':8,
-        'HST':-10,
-        'ICT':7,
-        'IDT':3,
-        'IrishST':1,
-        'IsraelST':2,
-        'JST':9,
-        'KOST':11,
-        'LINT':4,
-        'MDT':-6,
-        'MHT':12,
-        'MSK':3,
-        'MST':-7,
-        'MYT':8,
-        'NUT':-11,
-        'NZDT':13,
-        'NZST':12,
-        'PDT':-7,
-        'PET':-5,
-        'PGT':10,
-        'PHT':8,
-        'PONT':11,
-        'PST':-8,
-        'SAST':2,
-        'SBT':11,
-        'SGT':8,
-        'SRT':-3,
-        'SST':-11,
-        'TAHT':-10,
-        'TLT':9,
-        'TVT':12,
-        'ULAST':9,
-        'ULAT':8,
-        'UYST':-2,
-        'UYT':-3,
-        'VET':-4,
-        'WAST':2,
-        'WAT':1,
-        'WEST':1,
-        'WET':0,
-        'WIB':7,
-        'WIT':9,
-        'WITA':8}
-
-TARGET_YEAR = ["yyyy", "yy", "YYYY", "YY", "Y", "y"]
-TARGET_MONTH = ["MM", "M", "MMM", "MMMMM"]
-TEXT_MONTHS = [("Jan", "January"),
-              ("Feb", "February"),      # TODO: "Febr"
-              ("Mar", "March"),
-              ("Apr", "April"),
-              ("May", "May"),
-              ("Jun", "June"),
-              ("Jul", "July"),
-              ("Aug", "August"),
-              ("Sep", "September"),
-              ("Oct", "October"),
-              ("Nov", "November"),
-              ("Dec", "December")]
-TARGET_DAY = ["dd", "d", "DD", "D"]
-TARGET_HOUR = ["hh", "h", "HH", "H"]
-TARGET_MINUTE = ["mm", "m"]
-TARGET_SECOND = ["ss", "s", "SS", "S"]
-TARGET_WEEKDAY = ["eee", "EEE", "eeeee", "EEEEE"]
-TEXT_WEEKDAYS = [("Mon", "Monday"),
-                ("Tue", "Tuesday"),     # TODO: "Tues"
-                ("Wed", "Wednesday"),
-                ("Thu", "Thursday"),    # TODO: "Thurs"
-                ("Fri", "Friday"),
-                ("Sat", "Saturday"),
-                ("Sun", "Sunday")]
+from .clean_date_utils import JUMP, WEEKDAYS, MONTHS, HMS, AM, PM, ZONE, \
+                             TARGET_YEAR, TARGET_MONTH, TARGET_DAY, \
+                             TARGET_HOUR, TARGET_MINUTE, TARGET_SECOND, \
+                             TEXT_MONTHS, TARGET_WEEKDAY, TEXT_WEEKDAYS
+from .clean_date_utils import ParsedDate, ParsedTargetFormat
 
 STATS = {"cleaned": 0, "null": 0, "unknown": 0}
-
-class ParsedDate():
-
-    """Attributes of a parsed date.
-    Attributes:
-        year: Value of year.
-        month: Value of month.
-        day: Value of day.
-        hour: Value of hour.
-        minute: Value of minute.
-        second: Value of second.
-        weekday: Value of weekday.
-        timezone: Value of timezone.
-        utc_offset_hours: Hours of timezone offset.
-        utc_offset_minutes: Minutes of timezone offset.
-        utc_offset_add: Timezone offset is pos or neg.
-        valid: if parsed values are all valid.
-    """
-
-    def __init__(self):
-        """
-        This function initiate parse_date
-        """
-        self.ymd = {'year': None,
-                    'month': None,
-                    'day': None}
-        self.hms = {'hour': None,
-                    'minute': None,
-                    'second': None}
-        self.weekday = None
-        self.tzinfo = {'timezone': None,
-                       'utc_add': None,
-                       'utc_offset_hours': None,
-                       'utc_offset_minutes': None}
-        self.valid = 'cleaned'
-
-    def set_year(self, year):
-        """
-        This function set value of year
-        Parameters
-        ----------
-        year
-            year value
-        """
-        if 1700 <= year <= 2500:
-            self.ymd['year'] = year
-            return True
-        self.valid = 'unknown'
-        return False
-
-    def set_month(self, month):
-        """
-        This function set value of month
-        Parameters
-        ----------
-        month
-            month value
-        """
-        if 1 <= month <= 12:
-            self.ymd['month'] = month
-            return True
-        self.valid = 'unknown'
-        return False
-
-    def set_day(self, day):
-        """
-        This function set value of day
-        Parameters
-        ----------
-        day
-            day value
-        """
-        if self.ymd['month'] in [1, 3, 5, 7, 8, 10, 12]:
-            if 1 <= day <= 31:
-                self.ymd['day'] = day
-                return True
-        if self.ymd['month'] in [4, 6, 9, 11]:
-            if 1 <= day <= 30:
-                self.ymd['day'] = day
-                return True
-        if self.ymd['month'] in [2]:
-            if self._is_leap_year():
-                if 1 <= day <= 29:
-                    self.ymd['day'] = day
-                    return True
-            else:
-                if 1 <= day <= 28:
-                    self.ymd['day'] = day
-                    return True
-        self.valid = 'unknown'
-        return False
-
-    def set_hour(self, hour):
-        """
-        This function set value of hour
-        Parameters
-        ----------
-        hour
-            hour value
-        """
-        if 0 <= hour < 24:
-            self.hms['hour'] = hour
-            return True
-        self.valid = 'unknown'
-        return False
-
-    def set_minute(self, minute):
-        """
-        This function set value of minute
-        Parameters
-        ----------
-        minute
-            minute value
-        """
-        if 0 <= minute < 60:
-            self.hms['minute'] = minute
-            return True
-        self.valid = 'unknown'
-        return False
-
-    def set_second(self, second):
-        """
-        This function set value of second
-        Parameters
-        ----------
-        second
-            second value
-        """
-        if 0 <= second < 60:
-            self.hms['second'] = second
-            return True
-        self.valid = 'unknown'
-        return False
-
-    def set_tzinfo(self,
-                   timezone = None,
-                   utc_add = None,
-                   utc_offset_hours = None,
-                   utc_offset_minutes = None):
-        """
-        This function set timezone info
-        Parameters
-        ----------
-        timezone
-            timezone value
-        utc_add
-            the offset is positive or negtive comaring to UTC
-        utc_offset_hours
-            value of offset hours
-        utc_offset_minutes
-            value of offset minutes
-        """
-        if not timezone is None:
-            if timezone in all_timezones or timezone in ZONE:
-                self.tzinfo['timezone'] = timezone
-                return True
-            self.valid = 'unknown'
-            return False
-        if not utc_add is None:
-            self.tzinfo['utc_add'] = utc_add
-        if not utc_offset_hours is None:
-            self.tzinfo['utc_offset_hours'] = utc_offset_hours
-        if not utc_offset_minutes is None:
-            self.tzinfo['utc_offset_minutes'] = utc_offset_minutes
-        return True
-
-    def set_weekday(self, weekday):
-        """
-        This function set value of weekday
-        Parameters
-        ----------
-        weekday
-            weekday value
-        """
-        if 1 <= weekday <= 7:
-            self.weekday = weekday
-            return True
-        self.valid = 'unknown'
-        return False
-
-    def _is_leap_year(self):
-        """
-        This function judge if year is leap year
-        """
-        if self.ymd['year'] % 4 == 0:
-            if self.ymd['year'] % 100 == 0:
-                return self.ymd['year'] % 400 == 0
-            else:
-                return True
-        else:
-            return False
-
-class ParsedTargetFormat():
-
-    """Attributes of a parsed target format.
-    Attributes:
-        year_token: Token standing of year.
-        month_token: Token standing of month.
-        day_token: Token standing of day.
-        hour_token: Token standing of hour.
-        minute_token: Token standing of minute.
-        second_token: Token standing of second.
-        weekday_token: Token standing of weekday.
-        timezone_token: Token standing of timezone.
-        timezone: Value of target timezone.
-        utc_offset_hours: Hours of timezone offset.
-        utc_offset_minutes: Minutes of timezone offset.
-        utc_offset_add: Timezone offset is pos or neg.
-        valid: if target format is valid.
-        invalid_tokens: if target format is not valid,
-                        what tokens lead to this result.
-    """
-
-    def __init__(self):
-        """
-        This function initiate parsed_target_fomat
-        """
-        self.ymd_token = {'year_token': None,
-                          'month_token': None,
-                          'day_token': None}
-        self.hms_token = {'hour_token': None,
-                          'minute_token': None,
-                          'second_token': None,
-                          'ispm': False}
-        self.weekday_token = None
-        self.timezone_token = None
-        self.tzinfo = {'timezone': None,
-                       'utc_add': None,
-                       'utc_offset_hours': None,
-                       'utc_offset_minutes': None}
-        self.valid = True
-        self.invalid_tokens = []
-
-    def set_year_token(self, year_token):
-        """
-        This function set value of year_token
-        Parameters
-        ----------
-        year_token
-            token string of year
-        """
-        self.ymd_token['year_token'] = year_token
-        return True
-
-    def set_month_token(self, month_token):
-        """
-        This function set value of month_token
-        Parameters
-        ----------
-        month_token
-            token string of month
-        """
-        self.ymd_token['month_token'] = month_token
-        return True
-
-    def set_day_token(self, day_token):
-        """
-        This function set value of day_token
-        Parameters
-        ----------
-        day_token
-            token string of day
-        """
-        self.ymd_token['day_token'] = day_token
-        return True
-
-    def set_hour_token(self, hour_token):
-        """
-        This function set value of hour_token
-        Parameters
-        ----------
-        hour_token
-            token string of hour
-        """
-        self.hms_token['hour_token'] = hour_token
-        return True
-
-    def set_minute_token(self, minute_token):
-        """
-        This function set value of minute_token
-        Parameters
-        ----------
-        minute_token
-            token string of minute
-        """
-        self.hms_token['minute_token'] = minute_token
-        return True
-
-    def set_second_token(self, second_token):
-        """
-        This function set value of second_token
-        Parameters
-        ----------
-        second_token
-            token string of second
-        """
-        self.hms_token['second_token'] = second_token
-        return True
-
-    def set_weekday_token(self, weekday_token):
-        """
-        This function set value of weekday_token
-        Parameters
-        ----------
-        weekday_token
-            token string of weekday
-        """
-        self.weekday_token = weekday_token
-        return True
-
-    def set_timezone_token(self, timezone_token):
-        """
-        This function set value of timezone_token
-        Parameters
-        ----------
-        timezone_token
-            token string of timezone
-        """
-        self.timezone_token = timezone_token
-        return True
-
-    def set_tzinfo(self,
-                   timezone = None,
-                   utc_add = None,
-                   utc_offset_hours = None,
-                   utc_offset_minutes = None):
-        """
-        This function set timezone info
-        Parameters
-        ----------
-        timezone
-            name of timezone
-        utc_add
-            the offset is positive or negtive comaring to UTC
-        utc_offset_hours
-            value of offset hours
-        utc_offset_minutes
-            value of offset minutes
-        """
-        if not timezone is None:
-            self.tzinfo['timezone'] = timezone
-        if not utc_add is None:
-            self.tzinfo['utc_add'] = utc_add
-        if not utc_offset_hours is None:
-            self.tzinfo['utc_offset_hours'] = utc_offset_hours
-        if not utc_offset_minutes is None:
-            self.tzinfo['utc_offset_minutes'] = utc_offset_minutes
-        return True
-
-    def set_valid(self, valid):
-        """
-        This function set valid status of target format
-        Parameters
-        ----------
-        valid
-            valid status
-        """
-        self.valid = valid
-        return True
-
-    def set_ispm(self, ispm):
-        """
-        This function set value of judgement
-        of PM status for target format
-        Parameters
-        ----------
-        ispm
-            If is PM, True. If not, False
-        """
-        self.hms_token['ispm'] = ispm
-        return True
-
-    def add_invalid_token(self, token):
-        """
-        This function set value of invalid tokens
-        in target format
-        Parameters
-        ----------
-        token
-            invalid token
-        """
-        self.invalid_tokens.append(token)
-        return True
-
 
 def clean_date(
     df: Union[pd.DataFrame, dd.DataFrame],
@@ -596,24 +72,16 @@ def clean_date(
             f'fix_empty {fix_empty} is invalid. '
             f'It needs to be "auto_minimum", "auto_nearest" or "empty"'
         )
-
     if origin_timezone not in all_timezones and origin_timezone not in ZONE:
-        raise ValueError(
-            f'origin_timezone {origin_timezone} doesn\'t exist'
-        )
-
+        raise ValueError(f'origin_timezone {origin_timezone} doesn\'t exist')
     if target_timezone not in all_timezones and \
        target_timezone not in ZONE and \
        not target_timezone is None:
-        raise ValueError(
-            f'target_timezone {target_timezone} doesn\'t exist'
-        )
-
+        raise ValueError(f'target_timezone {target_timezone} doesn\'t exist')
     df = to_dask(df)
     # specify the metadata for dask apply
     meta = df.dtypes.to_dict()
     meta[f"{col}_clean"] = str
-
     df = df.apply(
         format_date,
         args=(col,
@@ -623,15 +91,11 @@ def clean_date(
         axis=1,
         meta=meta,
     )
-
     df, nrows = dask.compute(df, df.shape[0])
-
     # output the report describing the changes to the column
     if show_report:
         create_report("Date", STATS, nrows)
-
     return df
-
 
 def format_date(
     row: pd.Series,
@@ -644,36 +108,12 @@ def format_date(
     This function cleans date string.
     Parameters
     ----------
-    df
-        Pandas or Dask DataFrame.
-    col
-        Column name containing phone numbers.
-    target_format
-        The desired format of the date.
-        Defalut value is 'YYYY-MM-DD hh:mm:ss'
-    tz_info
-        Timezone of origin data and target data
-    fix_empty
-        The user can specify the way of fixing empty value from value set:
-            {'empty', 'auto_nearest', 'auto_minimum'}.
-        The default fixed_empty is "auto_minimum":
-            For hours, minutes and seconds:
-                Just fill them with zeros.
-            For years, months and days:
-                Fill it with the minimum value.
-        "auto_nearest":
-            For hours, minutes and seconds:
-                Fill them with nearest hour, minutes and seconds.
-            For years, months and days:
-                Fill it with the nearest value.
-        "empty":
-            Just left the missing component as it is
+    df, col, target_format, tz_info, fix_empty
+        same as explained in clean_date function
     """
-
     date = row[col]
     origin_timezone = tz_info['origin_tz']
     target_timezone = tz_info['target_tz']
-
     if check_date(date) == 'null':
         STATS['null'] += 1
         row[f"{col}_clean"] = np.nan
@@ -685,7 +125,6 @@ def format_date(
         parsed_date_data = parse(date=date, fix_empty=fix_empty)
         parsed_date_data.set_tzinfo(timezone=origin_timezone)
         parsed_date_data = set_timezone_offset(origin_timezone, parsed_date_data)
-
         # Handle target format and timezone
         parsed_target_format_data = check_target_format(target_format=target_format)
         parsed_target_format_data = set_timezone_offset(target_timezone, parsed_target_format_data)
@@ -707,14 +146,12 @@ def format_date(
                 f'Invalid tokens are {parsed_target_format_data.invalid_tokens}. '
                 f'Please retype it.'
             )
-
     return row
 
 def set_timezone_offset(timezone: Union[str, Any],
                         parsed_data: Union[ParsedDate, ParsedTargetFormat, Any]) -> Any:
     """
-    This function set timezone information for
-    parsed date or parsed target format
+    This function set timezone information for parsed date or parsed target format
     Parameters
     ----------
     timezone
@@ -768,12 +205,10 @@ def check_date(date: Union[str, Any]) -> Any:
         return 'null'
     tokes = split(date, JUMP)
     remain_tokens = tokes.copy()
-
     # Handle timezone
     for token in tokes:
         if token in all_timezones or token in ZONE:
             remain_tokens.remove(token)
-
     for token in tokes:
         if token in MONTHS or token in WEEKDAYS:
             remain_tokens.remove(token)
@@ -795,7 +230,6 @@ def validate_target_format(target_format: Union[str, Any]) -> Any:
     target_format
         target_format string
     """
-
     return check_target_format(target_format)
 
 def check_target_format(target_format: Union[str, Any]) -> Any:
@@ -809,29 +243,23 @@ def check_target_format(target_format: Union[str, Any]) -> Any:
     result = ParsedTargetFormat()
     target_tokens = split(target_format, JUMP)
     remain_tokens = deepcopy(target_tokens)
-
     # Handle Timezone
     result, remain_tokens = \
         figure_target_format_timezone(result, target_tokens, remain_tokens)
-
     # Handle year, month, day
     result, remain_tokens = \
         figure_target_format_ymd(result, target_tokens, remain_tokens)
-
     # Handle AM, PM with JUMP seperators
     result, remain_tokens = \
         figure_target_format_ampm(result, target_tokens, remain_tokens)
-
     # Handle hour, minute, second
     result, remain_tokens = \
         figure_target_format_hms(result, remain_tokens)
-
     # If len(remain_tokens) = 0, then is valid format
     if len(remain_tokens) > 0:
         result.set_valid(False)
         for token in remain_tokens:
             result.add_invalid_token(token)
-
     return result
 
 def figure_target_format_timezone(parsed_data: Union[ParsedTargetFormat, Any],
@@ -852,12 +280,10 @@ def figure_target_format_timezone(parsed_data: Union[ParsedTargetFormat, Any],
         if token in all_timezones:
             parsed_data.set_tzinfo(timezone = token)
             remain_tokens.remove(token)
-
     for token in target_tokens:
         if token in ('z', 'Z'):
             parsed_data.set_timezone_token(token)
             remain_tokens.remove(token)
-
     return parsed_data, remain_tokens
 
 def figure_target_format_ymd(parsed_data: Union[ParsedTargetFormat, Any],
@@ -887,7 +313,6 @@ def figure_target_format_ymd(parsed_data: Union[ParsedTargetFormat, Any],
         if token in TARGET_WEEKDAY:
             parsed_data.set_weekday_token(token)
             remain_tokens.remove(token)
-
     return parsed_data, remain_tokens
 
 def figure_target_format_ampm(parsed_data: Union[ParsedTargetFormat, Any],
@@ -910,11 +335,10 @@ def figure_target_format_ampm(parsed_data: Union[ParsedTargetFormat, Any],
         if token in PM:
             parsed_data.set_ispm(True)
             remain_tokens.remove(token)
-
     return parsed_data, remain_tokens
 
 def figure_target_format_hms(parsed_data: Union[ParsedTargetFormat, Any],
-                           remain_tokens: Union[list, Any]) -> Any:
+                             remain_tokens: Union[list, Any]) -> Any:
     """
     This function figure hour, minute and second token in target format
     Parameters
@@ -930,9 +354,7 @@ def figure_target_format_hms(parsed_data: Union[ParsedTargetFormat, Any],
             if not token in TARGET_MONTH and not token in TARGET_WEEKDAY and \
                not token in AM and not token in PM:
                 remain_str = token
-
         parsed_data, hms_tokens = get_target_format_hms_tokens(parsed_data, remain_str)
-
         for token in hms_tokens:
             if token in TARGET_HOUR:
                 parsed_data.set_hour_token(token)
@@ -942,7 +364,6 @@ def figure_target_format_hms(parsed_data: Union[ParsedTargetFormat, Any],
                 parsed_data.set_second_token(token)
         if len(remain_str) > 0:
             remain_tokens.remove(remain_str)
-
     return parsed_data, remain_tokens
 
 def get_target_format_hms_tokens(parsed_data: Union[ParsedTargetFormat, Any],
@@ -956,7 +377,6 @@ def get_target_format_hms_tokens(parsed_data: Union[ParsedTargetFormat, Any],
     remain_str
         remained string after figuring tokens
     """
-
     if 'z' in remain_str:
         parsed_data.timezone_token = 'z'
         hms_tokens = split(remain_str, [":", parsed_data.timezone_token])
@@ -976,7 +396,6 @@ def get_target_format_hms_tokens(parsed_data: Union[ParsedTargetFormat, Any],
             break
     if len(hms_tokens) == 0:
         hms_tokens = split(remain_str, [":"])
-
     return parsed_data, hms_tokens
 
 def ensure_ymd(tokes: Union[str, Any]) -> Any:
@@ -990,7 +409,6 @@ def ensure_ymd(tokes: Union[str, Any]) -> Any:
     result = ParsedDate()
     remain_tokens = tokes.copy()
     result, remain_tokens = ensure_year(result, tokes, remain_tokens)
-
     if len(remain_tokens) == 0:
         return result, remain_tokens
     num_tokens = []
@@ -1003,7 +421,6 @@ def ensure_ymd(tokes: Union[str, Any]) -> Any:
         result = ensure_month_day(result, num_tokens)
     else:
         result = ensure_year_month_day(result, num_tokens)
-
     return result, remain_tokens
 
 def ensure_year(parsed_data: Union[ParsedDate, Any],
@@ -1020,7 +437,6 @@ def ensure_year(parsed_data: Union[ParsedDate, Any],
     remain_tokens
         remained tokens
     """
-
     for token in tokes:
         if token in MONTHS:
             parsed_data.set_month(MONTHS[token])
@@ -1033,7 +449,6 @@ def ensure_year(parsed_data: Union[ParsedDate, Any],
             parsed_data.set_year(int(token))
             remain_tokens.remove(token)
             break
-
     return parsed_data, remain_tokens
 
 def ensure_month_day(parsed_data: Union[ParsedDate, Any],
@@ -1047,7 +462,6 @@ def ensure_month_day(parsed_data: Union[ParsedDate, Any],
     num_tokens
         remained numerical tokens
     """
-
     if len(num_tokens) == 1:
         if not parsed_data.month is None:
             parsed_data.set_day(int(num_tokens[0]))
@@ -1063,7 +477,6 @@ def ensure_month_day(parsed_data: Union[ParsedDate, Any],
         else:
             parsed_data.set_month(int(num_tokens[0]))
             parsed_data.set_day(int(num_tokens[1]))
-
     return parsed_data
 
 def ensure_year_month_day(parsed_data: Union[ParsedDate, Any],
@@ -1077,7 +490,6 @@ def ensure_year_month_day(parsed_data: Union[ParsedDate, Any],
     num_tokens
         remained numerical tokens
     """
-
     if len(num_tokens) == 1:
         parsed_data.set_year(int(num_tokens[-1]) + 2000)
     elif len(num_tokens) == 2:
@@ -1097,7 +509,6 @@ def ensure_year_month_day(parsed_data: Union[ParsedDate, Any],
         else:
             parsed_data.set_month(int(num_tokens[0]))
             parsed_data.set_day(int(num_tokens[1]))
-
     return parsed_data
 
 def ensure_hms(inner_result: Union[ParsedDate, Any],
@@ -1114,7 +525,6 @@ def ensure_hms(inner_result: Union[ParsedDate, Any],
     result = deepcopy(inner_result)
     remain_str = remain_tokens[0]
     hms_tokens = []
-
     # Judge the expression of am pm
     ispm = False
     for token in AM:
@@ -1151,7 +561,6 @@ def ensure_pm(parsed_data: Union[ParsedDate, Any],
         if it is pm time, offset = 12
         otherwise, offset = 0
     """
-
     if len(hms_tokens) == 1:
         parsed_data.set_hour(int(hms_tokens[0]) + offset)
     elif len(hms_tokens) == 2:
@@ -1206,7 +615,6 @@ def fix_empty_auto_nearest(parsed_res: Union[ParsedDate, Any]) -> Any:
     parsed_res
         parsed date result
     """
-
     now_time = datetime.datetime.now()
     if parsed_res.year is None:
         parsed_res.set_year(now_time.year)
@@ -1223,7 +631,6 @@ def fix_empty_auto_nearest(parsed_res: Union[ParsedDate, Any]) -> Any:
     if parsed_res.weekday is None:
         temp_date = datetime.datetime(parsed_res.year, parsed_res.month, parsed_res.day)
         parsed_res.set_weekday(temp_date.weekday() + 1)
-
     return parsed_res
 
 def fix_empty_auto_minimum(parsed_res: Union[ParsedDate, Any]) -> Any:
@@ -1234,7 +641,6 @@ def fix_empty_auto_minimum(parsed_res: Union[ParsedDate, Any]) -> Any:
     parsed_res
         parsed date result
     """
-
     if parsed_res.year is None:
         parsed_res.set_year(2000)
     if parsed_res.month is None:
@@ -1250,7 +656,6 @@ def fix_empty_auto_minimum(parsed_res: Union[ParsedDate, Any]) -> Any:
     if parsed_res.weekday is None:
         temp_date = datetime.datetime(parsed_res.year, parsed_res.month, parsed_res.day)
         parsed_res.set_weekday(temp_date.weekday() + 1)
-
     return parsed_res
 
 def parse(date: Union[str, Any], fix_empty: Union[str, Any]) -> Any:
@@ -1285,12 +690,10 @@ def change_timezone(parsed_date_data: Union[ParsedDate, Any],
     """
     origin_tz_offset = None
     target_tz_offset = None
-    origin_date = datetime.datetime(year=parsed_date_data.year,
-                                    month=parsed_date_data.month,
-                                    day=parsed_date_data.day,
-                                    hour=parsed_date_data.hour,
-                                    minute=parsed_date_data.minute,
-                                    second=parsed_date_data.second)
+    origin_date = \
+        datetime.datetime(year=parsed_date_data.year, month=parsed_date_data.month,
+                          day=parsed_date_data.day, hour=parsed_date_data.hour,
+                          minute=parsed_date_data.minute, second=parsed_date_data.second)
     if parsed_date_data.timezone in all_timezones:
         origin_tz_offset = pytz.timezone(parsed_date_data.timezone).utcoffset(origin_date)
         origin_tz_offset = timedelta(days=-origin_tz_offset.days,
@@ -1301,7 +704,6 @@ def change_timezone(parsed_date_data: Union[ParsedDate, Any],
         target_tz_offset = pytz.timezone(target_timezone).utcoffset(origin_date)
     elif target_timezone in ZONE:
         target_tz_offset = timedelta(seconds=ZONE[target_timezone] * 3600)
-
     result = deepcopy(parsed_date_data)
     if None in [parsed_date_data.year, parsed_date_data.mont,
                 parsed_date_data.day, parsed_date_data.hour,
@@ -1358,7 +760,6 @@ def transform_year(result_str: Union[str, Any],
                     result = result.replace(year_token, "0" + str(year))
                 else:
                     result = result.replace(year_token, str(year))
-
     return result
 
 def transform_month(result_str: Union[str, Any],
@@ -1398,7 +799,6 @@ def transform_month(result_str: Union[str, Any],
                 result = result.replace(month_token, TEXT_MONTHS[month - 1][1], 1)
             else:
                 result = result.replace(month_token, str(month), 1)
-
     return result
 
 def transform_day(result_str: Union[str, Any],
@@ -1430,13 +830,12 @@ def transform_day(result_str: Union[str, Any],
                     result = result.replace(day_token, str(day), 1)
             else:
                 result = result.replace(day_token, str(day))
-
     return result
 
-def transform_hour(result_str: Union[str, Any],
-                   hour_token: Union[str, Any],
+def transform_hms(result_str: Union[str, Any],
+                   hms_token: Union[str, Any],
                    ispm: Union[bool, Any],
-                   hour: Union[int, Any]) -> Any:
+                   hms_value: Union[int, Any]) -> Any:
     """
     This function transform parsed month into target format
     Parameters
@@ -1449,87 +848,22 @@ def transform_hour(result_str: Union[str, Any],
         value of hour
     """
     result = deepcopy(result_str)
-    if not hour_token is None:
-        if hour is None:
-            if len(hour_token) == 2:
-                result = result.replace(hour_token, str('--'))
-            elif len(hour_token) == 1:
-                result = result.replace(hour_token, str('-'))
+    if not hms_token is None:
+        if hms_value is None:
+            if len(hms_token) == 2:
+                result = result.replace(hms_token, str('--'))
+            elif len(hms_token) == 1:
+                result = result.replace(hms_token, str('-'))
         else:
-            if ispm:
-                hour = hour - 12
-            if len(hour_token) == 2:
-                if hour < 10:
-                    result = result.replace(hour_token, "0" + str(hour), 1)
+            if not ispm is None and ispm:
+                hms_value = hms_value - 12
+            if len(hms_token) == 2:
+                if hms_value < 10:
+                    result = result.replace(hms_token, "0" + str(hms_value), 1)
                 else:
-                    result = result.replace(hour_token, str(hour), 1)
+                    result = result.replace(hms_token, str(hms_value), 1)
             else:
-                result = result.replace(hour_token, str(hour))
-
-    return result
-
-def transform_minute(result_str: Union[str, Any],
-                   minute_token: Union[str, Any],
-                   minute: Union[int, Any]) -> Any:
-    """
-    This function transform parsed month into target format
-    Parameters
-    ----------
-    result_str
-        result string
-    minute_token
-        token of minute
-    minute
-        value of minute
-    """
-    result = deepcopy(result_str)
-    if not minute_token is None:
-        if minute is None:
-            if len(minute_token) == 2:
-                result = result.replace(minute_token, str('--'))
-            elif len(minute_token) == 1:
-                result = result.replace(minute_token, str('-'))
-        else:
-            if len(minute_token) == 2:
-                if minute < 10:
-                    result = result.replace(minute_token, "0" + str(minute), 1)
-                else:
-                    result = result.replace(minute_token, str(minute), 1)
-            else:
-                result = result.replace(minute_token, str(minute))
-
-    return result
-
-def transform_second(result_str: Union[str, Any],
-                   second_token: Union[str, Any],
-                   second: Union[int, Any]) -> Any:
-    """
-    This function transform parsed month into target format
-    Parameters
-    ----------
-    result_str
-        result string
-    second_token
-        token of second
-    second
-        value of second
-    """
-    result = deepcopy(result_str)
-    if not second_token is None:
-        if second is None:
-            if len(second_token) == 2:
-                result = result.replace(second_token, str('--'))
-            elif len(second_token) == 1:
-                result = result.replace(second_token, str('-'))
-        else:
-            if len(second_token) == 2:
-                if second < 10:
-                    result = result.replace(second_token, "0" + str(second), 1)
-                else:
-                    result = result.replace(second_token, str(second), 1)
-            else:
-                result = result.replace(second_token, str(second))
-
+                result = result.replace(hms_token, str(hms_value))
     return result
 
 def transform_weekday(result_str: Union[str, Any],
@@ -1558,7 +892,6 @@ def transform_weekday(result_str: Union[str, Any],
                 result = result.replace(weekday_token, TEXT_WEEKDAYS[weekday - 1][0])
             elif len(weekday_token) == 5:
                 result = result.replace(weekday_token, TEXT_WEEKDAYS[weekday - 1][1])
-
     return result
 
 def transform_timezone(result_str: Union[str, Any],
@@ -1591,10 +924,7 @@ def transform_timezone(result_str: Union[str, Any],
             if len(offset_minutes_str) == 1:
                 offset_minutes_str = '0' + offset_minutes_str
             result = result.replace(timezone_token,
-                                    "UTC" + utc_add +
-                                    offset_hours_str + ":"
-                                    + offset_minutes_str)
-
+                                    "UTC" + utc_add + offset_hours_str + ":" + offset_minutes_str)
     return result
 
 def transform(parsed_date_data: Union[ParsedDate, Any],
@@ -1618,7 +948,6 @@ def transform(parsed_date_data: Union[ParsedDate, Any],
     result = deepcopy(target_format)
     if not target_timezone is None:
         parsed_date_data = change_timezone(parsed_date_data, target_timezone)
-
     # Handle year
     result = transform_year(result,
                             parsed_target_format_data.year_token,
@@ -1628,26 +957,28 @@ def transform(parsed_date_data: Union[ParsedDate, Any],
                            parsed_target_format_data.day_token,
                            parsed_date_data.day)
     # Handle hours
-    result = transform_hour(result,
-                            parsed_target_format_data.hour_token,
-                            parsed_target_format_data.ispm,
-                            parsed_date_data.hour)
+    result = transform_hms(result,
+                           parsed_target_format_data.hour_token,
+                           parsed_target_format_data.ispm,
+                           parsed_date_data.hour)
     # Handle minutes
-    result = transform_minute(result,
-                             parsed_target_format_data.minute_token,
-                             parsed_date_data.minute)
+    result = transform_hms(result,
+                           parsed_target_format_data.minute_token,
+                           None,
+                           parsed_date_data.minute)
     # Handle seconds
-    result = transform_second(result,
-                             parsed_target_format_data.second_token,
-                             parsed_date_data.second)
+    result = transform_hms(result,
+                           parsed_target_format_data.second_token,
+                           None,
+                           parsed_date_data.second)
     # Handle month
     result = transform_month(result,
                              parsed_target_format_data.month_token,
                              parsed_date_data.month)
     # Handle weekday
     result = transform_weekday(result,
-                             parsed_target_format_data.weekday_token,
-                             parsed_date_data.weekday)
+                               parsed_target_format_data.weekday_token,
+                               parsed_date_data.weekday)
     # Handle timezone
     tz_info = {'utc_add': parsed_date_data.utc_add,
                'utc_offset_hours': parsed_date_data.utc_offset_hours,
